+++ CHANGELOG.md
@@ -7,6 +7,14 @@ This change log follows the format documented in [Keep a CHANGELOG].
 [semantic versioning]: http://semver.org/
 [keep a changelog]: http://keepachangelog.com/
 
+## 1.2.0 - 2021-01-06
+
+### Added
+
+- Add `landing` property to the location type contains information on how the app gets landed at the location, such as `redirected.
+
+- Add `redirect` function to the router that automatically sets `redirected: true` to the `landing` property.
+
 ## 1.1.2 - 2020-06-04
 
 ### Fixed